+++ local-store/LocalStorageStore.js
@@ -9,7 +9,7 @@ var LocalStorageStore = /** @class */ (function () {
     /**
      * @param store the underlying data store, this could be a string based store (i.e. native browser 'localStorage' or a MemoryStore instance) or it could be another LocalStore.
      * Note: the optional getKeys() function MUST return a new array each time it is called
-     * @param getStoreKeys a function that gets the keys from the 'store' (Note: this function function MUST return a new array each time it is called)
+     * @param getStoreKeys a function that gets the keys from the 'store' (Note: this function MUST return a new array each time it is called)
      * @param handleFullStore the handler to call when 'store' fails to store an item
      * @param trackKeysAndLen true to track the number of items and item keys added to this store
      * @param trackTotalSize true to track the total data size of the items in this store
@@ -172,29 +172,17 @@ var LocalStorageStore = /** @class */ (function () {
             }
         }
     };
-    /**
-     * @param store the underlying data store, this could be a string based store (i.e. native browser 'localStorage' or a MemoryStore instance) or it could be another LocalStore.
-     * Note: the optional getKeys() function MUST return a new array each time it is called
-     * @param getStoreKeys a function that gets the keys from the 'store' (Note: this function MUST return a new array each time it is called)
-     * @param trackKeysAndLen true to track the number of items and item keys added to this store
-     * @param trackTotalSize true to track the total data size of the items in this store
-     * @param maxValueSizeBytes an optional maximum size of values stored in this store
-     * @param loadExistingData optional flag to enable filtering the keys from the 'store' and load those which match the 'keyFilter'
-     * @param keyFilter optional storage key filter used if 'loadExistingData'
-     */
-    LocalStorageStore.newInst = function (store, getStoreKeys, handleFullStore, trackKeysAndLen, trackTotalSize, maxValueSizeBytes, loadExistingData, keyFilter) {
-        if (maxValueSizeBytes === void 0) { maxValueSizeBytes = 1000000; }
-        return new LocalStorageStore(store, getStoreKeys, handleFullStore, trackKeysAndLen, trackTotalSize, maxValueSizeBytes, loadExistingData, keyFilter);
-    };
-    /** Create a LocalStore object from a StorageLike object and an optional item removal callback
+    /** Create a LocalStore object from a StorageLike object with additional options and handlers/callbacks
      * @param store the store that will be used to store data.
      * Note: the optional getKeys() function MUST return a new array each time it is called
      * @param itemsRemovedCallback optional callback to call when items are removed from the store to free up space
      * @param logInfo optional flag to log store clearing events to the key-value store
      * @param removeRatio optional percentage of items to remove from the store when it's full
+     * @param keyParser optional function which parses a 'store' key and extracts a numeric sort order value (default: parseInt)
      */
-    LocalStorageStore.newTimestampInst = function (store, itemsRemovedCallback, logInfo, removePercentage) {
-        var clearer = ClearFullStore.newInst(Number.parseInt, itemsRemovedCallback, removePercentage);
+    LocalStorageStore.newTimestampInst = function (store, itemsRemovedCallback, logInfo, removePercentage, keyParser) {
+        if (keyParser === void 0) { keyParser = parseInt; }
+        var clearer = ClearFullStore.newInst(keyParser, itemsRemovedCallback, removePercentage);
         return new LocalStorageStore(store, null, function (store, err) {
             clearer.clearOldItems(store, logInfo, err);
         }, true, true, undefined, true);