+++ local-store/LocalStoreByTimestamp.js
@@ -60,9 +60,11 @@ var LocalStoreByTimestamp = /** @class */ (function () {
      * @param logInfo optional flag to log store clearing events to the key-value store
      * @param removeRatio optional percentage of items to remove from the store when it's full
      * @param keyGenerator optional function which generates a key for storing items when addItem() is called, default is 'UniqueChronologicalKeys.uniqueTimestamp'
+     * @param keyParser optional function which parses a 'store' key and extracts a numeric sort order value (default: parseInt)
      */
-    LocalStoreByTimestamp.newTimestampInst = function (localStore, itemsRemovedCallback, logInfo, removePercentage, keyGenerator) {
-        var clearer = ClearFullStore.newInst(Number.parseInt, itemsRemovedCallback, removePercentage);
+    LocalStoreByTimestamp.newTimestampInst = function (localStore, itemsRemovedCallback, logInfo, removePercentage, keyGenerator, keyParser) {
+        if (keyParser === void 0) { keyParser = parseInt; }
+        var clearer = ClearFullStore.newInst(keyParser, itemsRemovedCallback, removePercentage);
         return new LocalStoreByTimestamp(localStore, keyGenerator || UniqueChronologicalKeys.uniqueTimestamp, function (storeInst, err) {
             clearer.clearOldItems(storeInst, logInfo, err);
         });