+++ test/LocalStoreByCategoryTest.ts
@@ -22,7 +22,7 @@ interface CategoryEvent {
 suite("LocalStoreByCategory", function LocalStoreByCategoryTest() {
 
     var memStore = MemoryStore.newInst();
-    var localStore = LocalStorageStore.newInst(memStore, null, <LocalStore.FullStoreHandler><any>null, true, false, 80, false);
+    var localStore = new LocalStorageStore(memStore, null, <LocalStore.FullStoreHandler><any>null, true, false, 80, false, undefined);
     var storeBldr = LocalStoreByCategory.Builder.newInst(localStore, UniqueChronologicalKeys.uniqueTimestamp);
     var store = storeBldr.addStores({
         alpha: storeBldr.toStore(BasicCategorizers.newPrefixCategorizer("alpha-")),
@@ -103,7 +103,7 @@ suite("LocalStoreByCategory", function LocalStoreByCategoryTest() {
         var cI = 0;
         var dI = 0;
         var memStore = MemoryStore.newInst(10);
-        var localStore = LocalStorageStore.newInst(memStore, null, <LocalStore.FullStoreHandler><any>null, true, false, 80, false);
+        var localStore = new LocalStorageStore(memStore, null, <LocalStore.FullStoreHandler><any>null, true, false, 80, false, undefined);
         var storeBldr = LocalStoreByCategory.Builder.newInst(localStore, UniqueChronologicalKeys.uniqueTimestamp);
         var store = storeBldr.addStores({
             charlie: storeBldr.toStore(BasicCategorizers.newPrefixCategorizer("c-"), (storeInst, removedItems, err, removedCount) => cI++),